requirements-dev.txt
pytest
pytest-django
factory_boy
pytest-factoryboy